+++ Profiler_Test/ProfilerTestBase.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Text;
 
 namespace Cqse.Teamscale.Profiler.Dotnet
 {
