+++ UploadDaemon/Configuration/AzureFileStorage.cs
@@ -1,6 +1,4 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
 
 namespace UploadDaemon.Configuration
 {
