+++ UploadDaemon/SymbolAnalysis/LineCoverageSynthesizer.cs
@@ -1,12 +1,10 @@
-﻿using UploadDaemon;
-using Cqse.ConQAT.Dotnet.Bummer;
+﻿using Cqse.ConQAT.Dotnet.Bummer;
 using NLog;
 using System;
 using System.Collections.Generic;
-using System.IO.Abstractions;
 using System.Linq;
 using System.Text;
-using UploadDaemon.Utils;
+using UploadDaemon.Configuration;
 
 namespace UploadDaemon.SymbolAnalysis
 {
