+++ UploadDaemon/SymbolAnalysis/ParsedTraceFile.cs
@@ -2,9 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 
 namespace UploadDaemon.SymbolAnalysis
 {
