+++ UploadDaemon/Upload/IUploadFactory.cs
@@ -1,5 +1,4 @@
-﻿using UploadDaemon;
-using System.IO.Abstractions;
+﻿using System.IO.Abstractions;
 using UploadDaemon.Configuration;
 
 namespace UploadDaemon.Upload
