+++ UploadDaemon_Test/Archiving/PurgeArchiveTaskTest.cs
@@ -1,5 +1,4 @@
-﻿using UploadDaemon;
-using Moq;
+﻿using Moq;
 using NUnit.Framework;
 using System;
 using UploadDaemon.Configuration;
