+++ UploadDaemon_Test/Configuration/ConfigParserTest.cs
@@ -1,8 +1,7 @@
 ﻿using NUnit.Framework;
 using System;
-using UploadDaemon.Configuration;
 
-namespace UploadDaemon
+namespace UploadDaemon.Configuration
 {
     [TestFixture]
     public class ConfigParserTest
