+++ UploadDaemon_Test/FileSystemMockingUtils.cs
@@ -1,10 +1,6 @@
 ﻿using Moq;
 using System;
-using System.Collections.Generic;
 using System.IO.Abstractions;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 /// <summary>
 /// Utility functions for mocking System.IO.Abstractions classes.
