+++ UploadDaemon_Test/SymbolAnalysis/ParsedTraceFileTest.cs
@@ -1,36 +1,36 @@
 ﻿using NUnit.Framework;
-using UploadDaemon;
-using System.Collections.Generic;
-using UploadDaemon.SymbolAnalysis;
 
-[TestFixture]
-public class ParsedTraceFileTest
+namespace UploadDaemon.SymbolAnalysis
 {
-    [Test]
-    public void MustSupportNewStyleMethodReferences()
+    [TestFixture]
+    public class ParsedTraceFileTest
     {
-        ParsedTraceFile trace = new ParsedTraceFile(new string[]
+        [Test]
+        public void MustSupportNewStyleMethodReferences()
         {
+            ParsedTraceFile trace = new ParsedTraceFile(new string[]
+            {
             "Assembly=ProfilerGUI:2 Version:1.0.0.0",
             $"Inlined=2:12345",
-        }, "cov.txt");
+            }, "cov.txt");
 
-        Assert.That(trace.CoveredMethods, Has.Count.EqualTo(1));
-        Assert.That(trace.CoveredMethods[0].Item1, Is.EqualTo("ProfilerGUI"));
-        Assert.That(trace.CoveredMethods[0].Item2, Is.EqualTo(12345));
-    }
+            Assert.That(trace.CoveredMethods, Has.Count.EqualTo(1));
+            Assert.That(trace.CoveredMethods[0].Item1, Is.EqualTo("ProfilerGUI"));
+            Assert.That(trace.CoveredMethods[0].Item2, Is.EqualTo(12345));
+        }
 
-    [Test]
-    public void MustSupportOldStyleMethodReferences()
-    {
-        ParsedTraceFile trace = new ParsedTraceFile(new string[]
+        [Test]
+        public void MustSupportOldStyleMethodReferences()
         {
+            ParsedTraceFile trace = new ParsedTraceFile(new string[]
+            {
             "Assembly=ProfilerGUI:2 Version:1.0.0.0",
             $"Inlined=2:9876:12345",
-        }, "cov.txt");
+            }, "cov.txt");
 
-        Assert.That(trace.CoveredMethods, Has.Count.EqualTo(1));
-        Assert.That(trace.CoveredMethods[0].Item1, Is.EqualTo("ProfilerGUI"));
-        Assert.That(trace.CoveredMethods[0].Item2, Is.EqualTo(12345));
+            Assert.That(trace.CoveredMethods, Has.Count.EqualTo(1));
+            Assert.That(trace.CoveredMethods[0].Item1, Is.EqualTo("ProfilerGUI"));
+            Assert.That(trace.CoveredMethods[0].Item2, Is.EqualTo(12345));
+        }
     }
 }
