+++ UploadDaemon_Test/UploadTaskTest.cs
@@ -10,18 +10,19 @@ using System.Threading.Tasks;
 using UploadDaemon.SymbolAnalysis;
 using UploadDaemon.Upload;
 using UploadDaemon.Configuration;
-using UploadDaemon.Utils;
 
-[TestFixture]
-public class UploadTaskTest
+namespace UploadDaemon
 {
-    private const string TraceDirectory = @"C:\users\public\traces";
-    private const string TraceDirectoryWithSpace = @"C:\users\user with spaces\traces";
-    private const string PdbDirectory = @"C:\pdbs";
-    private const string RevisionFile = @"C:\revision.txt";
-    private const string VersionAssembly = "VersionAssembly";
+    [TestFixture]
+    public class UploadTaskTest
+    {
+        private const string TraceDirectory = @"C:\users\public\traces";
+        private const string TraceDirectoryWithSpace = @"C:\users\user with spaces\traces";
+        private const string PdbDirectory = @"C:\pdbs";
+        private const string RevisionFile = @"C:\revision.txt";
+        private const string VersionAssembly = "VersionAssembly";
 
-    private static readonly Config config = Config.Read($@"
+        private static readonly Config config = Config.Read($@"
         match:
           - uploader:
               versionAssembly: {VersionAssembly}
@@ -30,7 +31,7 @@ public class UploadTaskTest
               targetdir: {TraceDirectory}
     ");
 
-    private static readonly Config lineCoverageConfig = Config.Read($@"
+        private static readonly Config lineCoverageConfig = Config.Read($@"
         match:
           - uploader:
               directory: C:\store
@@ -40,7 +41,7 @@ public class UploadTaskTest
               targetdir: {TraceDirectory}
     ");
 
-    private static readonly Config archiveLineCoverageConfig = Config.Read($@"
+        private static readonly Config archiveLineCoverageConfig = Config.Read($@"
         archiveLineCoverage: true
         match:
           - uploader:
@@ -52,58 +53,26 @@ public class UploadTaskTest
               targetdir: {TraceDirectory}
     ");
 
-    [Test]
-    public void TracesShouldBeArchivedAfterASuccessfulUpload()
+        [Test]
+        public void TracesShouldBeArchivedAfterASuccessfulUpload()
 
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe
 Inlined=1:33555646:100678050" },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(config);
+            new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(config);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"uploaded\coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"uploaded\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TracesShouldBeArchivedAfterASuccessfulLineCoverageUpload()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TracesShouldBeArchivedAfterASuccessfulLineCoverageUpload()
         {
-            { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
-Process=foo.exe
-Inlined=1:33555646:100678050" },
-            { RevisionFile, @"revision: 12345" },
-            { PdbDirectory, new MockDirectoryData() },
-        });
-
-        new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(lineCoverageConfig);
-
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"uploaded\coverage_1_1.txt");
-    }
-
-    [Test]
-    public void TracesShouldNotBeArchivedAfterAFailedUpload()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
-        {
-            { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
-Process=foo.exe
-Inlined=1:33555646:100678050" },
-        });
-
-        new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(config);
-
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"coverage_1_1.txt");
-    }
-
-    [Test]
-    public void TracesShouldNotBeArchivedAfterAFailedLineCoverageUpload()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe
@@ -112,15 +81,30 @@ Inlined=1:33555646:100678050" },
             { PdbDirectory, new MockDirectoryData() },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(lineCoverageConfig);
+            new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(lineCoverageConfig);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"uploaded\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestArchivingTracesThatProduceNoLineCoverage()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TracesShouldNotBeArchivedAfterAFailedUpload()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        {
+            { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
+Process=foo.exe
+Inlined=1:33555646:100678050" },
+        });
+
+            new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(config);
+
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"coverage_1_1.txt");
+        }
+
+        [Test]
+        public void TracesShouldNotBeArchivedAfterAFailedLineCoverageUpload()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe
@@ -129,15 +113,15 @@ Inlined=1:33555646:100678050" },
             { PdbDirectory, new MockDirectoryData() },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer(false)).Run(lineCoverageConfig);
+            new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(lineCoverageConfig);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"no-line-coverage\coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestArchivingUploadedLineCoverage()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TestArchivingTracesThatProduceNoLineCoverage()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe
@@ -146,65 +130,82 @@ Inlined=1:33555646:100678050" },
             { PdbDirectory, new MockDirectoryData() },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(archiveLineCoverageConfig);
+            new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer(false)).Run(lineCoverageConfig);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"uploaded\coverage_1_1.txt", @"converted-line-coverage\coverage_1_1.txt.simple");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"no-line-coverage\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestArchivingTraceWithMissingVersion()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TestArchivingUploadedLineCoverage()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        {
+            { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
+Process=foo.exe
+Inlined=1:33555646:100678050" },
+            { RevisionFile, @"revision: 12345" },
+            { PdbDirectory, new MockDirectoryData() },
+        });
+
+            new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(archiveLineCoverageConfig);
+
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"uploaded\coverage_1_1.txt", @"converted-line-coverage\coverage_1_1.txt.simple");
+        }
+
+        [Test]
+        public void TestArchivingTraceWithMissingVersion()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=OtherAssembly:1 Version:4.0.0.0
 Process=foo.exe
 Inlined=1:33555646:100678050" },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(config);
+            new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(config);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"missing-version\coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"missing-version\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestArchivingTraceWithMissingProcess()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TestArchivingTraceWithMissingProcess()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=OtherAssembly:1 Version:4.0.0.0
 Inlined=1:33555646:100678050" },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(config);
+            new UploadTask(fileSystem, new MockUploadFactory(false), new MockLineCoverageSynthesizer()).Run(config);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"missing-process\coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"missing-process\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestArchivingEmptyTrace()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TestArchivingEmptyTrace()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectory("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe" },
         });
 
-        new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(config);
+            new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(config);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectory, @"empty-traces\coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectory, @"empty-traces\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestPathsWithSpaces()
-    {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+        [Test]
+        public void TestPathsWithSpaces()
+        {
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectoryWithSpace("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe
 Inlined=1:33555646:100678050" },
         });
 
-        Config configWithSpaceInTraceDirectory = Config.Read($@"
+            Config configWithSpaceInTraceDirectory = Config.Read($@"
             match:
               - uploader:
                   versionAssembly: {VersionAssembly}
@@ -213,15 +214,15 @@ Inlined=1:33555646:100678050" },
                   targetdir: {TraceDirectoryWithSpace}
         ");
 
-        new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(configWithSpaceInTraceDirectory);
+            new UploadTask(fileSystem, new MockUploadFactory(true), new MockLineCoverageSynthesizer()).Run(configWithSpaceInTraceDirectory);
 
-        AssertFilesInDirectory(fileSystem, TraceDirectoryWithSpace, @"uploaded\coverage_1_1.txt");
-    }
+            AssertFilesInDirectory(fileSystem, TraceDirectoryWithSpace, @"uploaded\coverage_1_1.txt");
+        }
 
-    [Test]
-    public void TestVersionPrefix()
-    {
-        Config config = Config.Read($@"
+        [Test]
+        public void TestVersionPrefix()
+        {
+            Config config = Config.Read($@"
             match:
               - uploader:
                   versionAssembly: {VersionAssembly}
@@ -231,86 +232,87 @@ Inlined=1:33555646:100678050" },
                   targetdir: {TraceDirectoryWithSpace}
         ");
 
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
             { FileInTraceDirectoryWithSpace("coverage_1_1.txt"), @"Assembly=VersionAssembly:1 Version:4.0.0.0
 Process=foo.exe
 Inlined=1:33555646:100678050" },
         });
 
-        MockUploadFactory uploadFactory = new MockUploadFactory(true);
-        new UploadTask(fileSystem, uploadFactory, new MockLineCoverageSynthesizer()).Run(config);
+            MockUploadFactory uploadFactory = new MockUploadFactory(true);
+            new UploadTask(fileSystem, uploadFactory, new MockLineCoverageSynthesizer()).Run(config);
 
-        uploadFactory.uploadMock.Verify(upload => upload.UploadAsync(It.IsAny<string>(), "prefix_4.0.0.0"));
-    }
-
-    private class MockUploadFactory : IUploadFactory
-    {
-        public readonly Mock<IUpload> uploadMock;
-
-        public MockUploadFactory(bool successfull)
-        {
-            this.uploadMock = new Mock<IUpload>();
-            uploadMock.Setup(upload => upload.UploadAsync(It.IsAny<string>(), It.IsAny<string>()))
-                .Returns(Task.FromResult(successfull));
-            uploadMock.Setup(upload => upload.UploadLineCoverageAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<RevisionFileUtils.RevisionOrTimestamp>()))
-                .Returns(Task.FromResult(successfull));
+            uploadFactory.uploadMock.Verify(upload => upload.UploadAsync(It.IsAny<string>(), "prefix_4.0.0.0"));
         }
 
-        /// <inheritdoc/>
-        public IUpload CreateUpload(Config.ConfigForProcess config, IFileSystem fileSystem)
+        private class MockUploadFactory : IUploadFactory
         {
-            return uploadMock.Object;
-        }
-    }
+            public readonly Mock<IUpload> uploadMock;
 
-    private class MockLineCoverageSynthesizer : ILineCoverageSynthesizer
-    {
-        private readonly bool shouldProduceCoverage;
-
-        public MockLineCoverageSynthesizer(bool shouldProduceCoverage = true)
-        {
-            this.shouldProduceCoverage = shouldProduceCoverage;
-        }
-
-        /// <inheritdoc/>
-        public Dictionary<string, FileCoverage> ConvertToLineCoverage(ParsedTraceFile traceFile, string symbolDirectory, GlobPatternList assemblyPatterns)
-        {
-            if (!shouldProduceCoverage)
+            public MockUploadFactory(bool successfull)
             {
-                return null;
+                this.uploadMock = new Mock<IUpload>();
+                uploadMock.Setup(upload => upload.UploadAsync(It.IsAny<string>(), It.IsAny<string>()))
+                    .Returns(Task.FromResult(successfull));
+                uploadMock.Setup(upload => upload.UploadLineCoverageAsync(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<RevisionFileUtils.RevisionOrTimestamp>()))
+                    .Returns(Task.FromResult(successfull));
             }
 
-            // Return some arbitrary coverage
-            FileCoverage fileCoverage = new FileCoverage();
-            fileCoverage.CoveredLineRanges.Add((12, 33));
-            return new Dictionary<string, FileCoverage>()
+            /// <inheritdoc/>
+            public IUpload CreateUpload(Config.ConfigForProcess config, IFileSystem fileSystem)
+            {
+                return uploadMock.Object;
+            }
+        }
+
+        private class MockLineCoverageSynthesizer : ILineCoverageSynthesizer
+        {
+            private readonly bool shouldProduceCoverage;
+
+            public MockLineCoverageSynthesizer(bool shouldProduceCoverage = true)
+            {
+                this.shouldProduceCoverage = shouldProduceCoverage;
+            }
+
+            /// <inheritdoc/>
+            public Dictionary<string, FileCoverage> ConvertToLineCoverage(ParsedTraceFile traceFile, string symbolDirectory, GlobPatternList assemblyPatterns)
+            {
+                if (!shouldProduceCoverage)
+                {
+                    return null;
+                }
+
+                // Return some arbitrary coverage
+                FileCoverage fileCoverage = new FileCoverage();
+                fileCoverage.CoveredLineRanges.Add((12, 33));
+                return new Dictionary<string, FileCoverage>()
             {
                 { "file1.cs", fileCoverage }
             };
+            }
         }
-    }
 
-    /// <summary>
-    /// Returns a file with the given name in the trace directory.
-    /// </summary>
-    private string FileInTraceDirectory(string fileName)
-    {
-        return Path.Combine(TraceDirectory, fileName);
-    }
+        /// <summary>
+        /// Returns a file with the given name in the trace directory.
+        /// </summary>
+        private string FileInTraceDirectory(string fileName)
+        {
+            return Path.Combine(TraceDirectory, fileName);
+        }
 
-    /// <summary>
-    /// Returns a file with the given name in the trace directory (which contains a space in the path).
-    /// </summary>
-    private string FileInTraceDirectoryWithSpace(string fileName)
-    {
-        return Path.Combine(TraceDirectoryWithSpace, fileName);
-    }
+        /// <summary>
+        /// Returns a file with the given name in the trace directory (which contains a space in the path).
+        /// </summary>
+        private string FileInTraceDirectoryWithSpace(string fileName)
+        {
+            return Path.Combine(TraceDirectoryWithSpace, fileName);
+        }
 
-    private void AssertFilesInDirectory(IFileSystem fileSystem, string directory, params string[] expectedFileNames)
-    {
-        string[] files = fileSystem.Directory.GetFiles(directory, "*.*", SearchOption.AllDirectories);
-        IEnumerable<string> relativePaths = files.Select(path => path.Substring(directory.Length + 1));
-        Assert.That(relativePaths, Is.EquivalentTo(expectedFileNames));
+        private void AssertFilesInDirectory(IFileSystem fileSystem, string directory, params string[] expectedFileNames)
+        {
+            string[] files = fileSystem.Directory.GetFiles(directory, "*.*", SearchOption.AllDirectories);
+            IEnumerable<string> relativePaths = files.Select(path => path.Substring(directory.Length + 1));
+            Assert.That(relativePaths, Is.EquivalentTo(expectedFileNames));
+        }
     }
 }
