+++ test-data/UploadDaemon.SymbolAnalysis.LineCoverageSynthesizerTest/ProfilerGUI.pdb
