+++ test-data/UploadDaemon.SymbolAnalysis.SymbolCollectionTest/Invalid.pdb
