+++ test-data/UploadDaemon.SymbolAnalysis.SymbolCollectionTest/ProfilerGUI.pdb
