+++ test-data/UploadDaemon.SymbolAnalysis.SymbolCollectionTest/ProfilerGUICopy.pdb
