+++ test-data/UploadDaemon.SymbolAnalysis.SymbolCollectionTest/sub/Sub.pdb
